refactor: Move HomePage to subfolder
