+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import HomePage from "@/components/HomePage";
+import HomePage from "@/components/HomePage/HomePage";
 import { SettingsProvider } from "@/context/SettingsContext";
 
 export default function Root() {
